+++ model_navigator/framework_api/contrib/huggingface/torch/export.py
@@ -45,6 +45,7 @@ from model_navigator.framework_api.pipelines.builders import (
     correctness_builder,
     preprocessing_builder,
     profiling_builder,
+    torch_conversion_builder,
     torch_export_builder,
 )
 from model_navigator.framework_api.pipelines.pipeline_manager import PipelineManager
@@ -236,7 +237,13 @@ def export(
         profiler_config=profiler_config,
     )
 
-    builders = [preprocessing_builder, torch_export_builder, correctness_builder, config_generation_builder]
+    builders = [
+        preprocessing_builder,
+        torch_export_builder,
+        torch_conversion_builder,
+        correctness_builder,
+        config_generation_builder,
+    ]
     if run_profiling:
         builders.append(profiling_builder)
     pipeline_manager = PipelineManager(builders)