+++ model_navigator/framework_api/tensorflow.py
@@ -25,6 +25,7 @@ from model_navigator.framework_api.pipelines.builders import (
     correctness_builder,
     preprocessing_builder,
     profiling_builder,
+    tensorflow_conversion_builder,
     tensorflow_export_builder,
 )
 from model_navigator.framework_api.pipelines.pipeline_manager import PipelineManager
@@ -64,7 +65,7 @@ def export(
     run_profiling: bool = True,
     profiler_config: Optional[ProfilerConfig] = None,
 ) -> PackageDescriptor:
-    """Function exports TensorFlow 2 model to all supported formats."""
+    """Function exports TensorFlow2 model to all supported formats."""
     if model_name is None:
         model_name = get_default_model_name()
     if workdir is None:
@@ -121,7 +122,13 @@ def export(
         profiler_config=profiler_config,
     )
 
-    builders = [preprocessing_builder, tensorflow_export_builder, correctness_builder, config_generation_builder]
+    builders = [
+        preprocessing_builder,
+        tensorflow_export_builder,
+        tensorflow_conversion_builder,
+        correctness_builder,
+        config_generation_builder,
+    ]
     if run_profiling:
         builders.append(profiling_builder)
     pipeline_manager = PipelineManager(builders)