+++ model_navigator/framework_api/torch.py
@@ -27,6 +27,7 @@ from model_navigator.framework_api.pipelines.builders import (
     correctness_builder,
     preprocessing_builder,
     profiling_builder,
+    torch_conversion_builder,
     torch_export_builder,
 )
 from model_navigator.framework_api.pipelines.pipeline_manager import PipelineManager
@@ -152,7 +153,13 @@ def export(
         profiler_config=profiler_config,
     )
 
-    builders = [preprocessing_builder, torch_export_builder, correctness_builder, config_generation_builder]
+    builders = [
+        preprocessing_builder,
+        torch_export_builder,
+        torch_conversion_builder,
+        correctness_builder,
+        config_generation_builder,
+    ]
     if run_profiling:
         builders.append(profiling_builder)
     pipeline_manager = PipelineManager(builders)